tc: add dualpi2 scheduler module
DUALPI2 AQM is a combination of the DUALQ Coupled-AQM with a PI2 base-AQM,
able to control scalable congestion controls like DCTCP and TCP-Prague.
The PI2 AQM details can be found in the paper:
CoNEXT '16 Proceedings of the 12th International on Conference on emerging
Networking EXperiments and Technologies : "PI2: A Linearized AQM for both
Classic and Scalable TCP".
PI2 makes quite some PIE heuristics unnecessary, allowing to use both
Reno/Cubic in parallel with DCTCP, maintaining window fairness.

Additional details can be found in the draft:
https://www.ietf.org/id/draft-ietf-tsvwg-aqm-dualq-coupled

Signed-off-by: Olga Albisser <olga@albisser.org>
Signed-off-by: Koen De Schepper <koen.de_schepper@nokia-bell-labs.com>
Signed-off-by: Oliver Tilmans <olivier.tilmans@nokia-bell-labs.com>
Signed-off-by: Bob Briscoe <research@bobbriscoe.net>
Signed-off-by: Henrik Steen <henrist@henrist.net>